Complete Choice Mode and delete note.

+++ app/src/main/java/com/wings/simplenote/model/INoteModel.java
@@ -12,7 +12,7 @@ public interface INoteModel {
 
     boolean updateNote(Note note);
 
-    void deleteNote(int id);
+    void deleteNote(Long id);
 
     List<Note> selectAll();
